GameMessageParentEvent refactor

+++ Source/ACE.Server/WorldObjects/Creature_Missile.cs
@@ -46,7 +46,7 @@ namespace ACE.Server.WorldObjects
                 delayChain.AddDelaySeconds(0.001f);     // ensuring this message gets sent after player broadcasts above...
                 delayChain.AddAction(this, () =>
                 {
-                    EnqueueBroadcast(new GameMessageParentEvent(this, ammo, (int)ACE.Entity.Enum.ParentLocation.RightHand, (int)ACE.Entity.Enum.Placement.RightHandCombat));
+                    EnqueueBroadcast(new GameMessageParentEvent(this, ammo, ACE.Entity.Enum.ParentLocation.RightHand, ACE.Entity.Enum.Placement.RightHandCombat));
                 });
                 delayChain.EnqueueChain();
             });
